pyproject.toml
[tool.poetry]
name = "transferapi"
version = "0.1.0"
description = ""
authors = ["churka <psinka2006@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
logging = "^0.4.9.6"
django-filter = "^25.1"
drf-yasg = "^1.21.9"
django-redis = "^5.4.0"
hiredis = "^3.1.0"
requests = "^2.32.3"
psutil = "^7.0.0"
pyTelegramBotAPI = "^4.26.0"
celery= "^5.5.0"
django-cors-headers = "^4.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"